[Refactor][Ops] Select a kernel and ask it how it is built
An op reached its kernels through several verbs, and each op wrote its own compile
boundary by hand. Both are now one path. `Op.kernel_for(role, inputs, call)` is the
single verb; `entry_for(role, call)` answers with the identity that keys the cache and
the builder that fills it. A role is a memoization bucket, one per kernel the op runs,
never the name of an implementation it chose.

Each op's `torch.library.custom_op` and its fake are generated from its manifest entry
by `_compile_boundary_codegen`. An op joins the boundary by declaring one
`OperatorSpec` per operator it registers; the entry says the rest, so the schema and
the declaration cannot drift apart and adding an op costs no registration code. The
last five registrations under `src/tileops/kernels/` are deleted with them: a boundary
in a kernel file makes the compiled graph depend on which kernel served the call.

An output's dtype comes from exactly one place, decided per output rather than per op.
The caller states it through a parameter named `out_dtype`, for the outputs the entry
marks `caller_stated: true`; the entry states every other one as a constant,
`same_as(x)` or `promote_int_to_float(x)`. A declaration resolves to one dtype and
never names a set, so an op returning an auxiliary tensor beside its result keeps that
one its own. `check_c9_output_dtype_convention` holds both directions, `_output_dtype`
is the single implementation the fake and the eager path share, and the compile
contract holds each fake to what its op returns.

`_fake_output_dtype` and `RoutingEpilogueSpec.output_dtype` are gone with the ambiguity
they covered for. `IndexedExpertMLPFwdOp` joins the manifest and reaches its kernels
and its boundary the same way every other op does. `TopkSelectorKernel` writes the
slots a narrow request leaves unselected, which it returned as whatever the buffer held.

+++ benchmarks/ops/bench_mamba.py
@@ -176,7 +176,7 @@ def test_da_cumsum_fwd_bench(
     op = DaCumsumFwdOp(
         chunk_len=chunk_len,
         dt_softplus=dt_softplus,
-        dtype=dtype,
+        out_dtype=dtype,
         tune=tune,
     )
     bm = ManifestBenchmark(op, test)
