+++ tests/ops/test_special_elementwise.py
@@ -258,7 +258,7 @@ class AlibiFixture(FixtureBase):
 def test_alibi(seq_len: int, num_heads: int, dtype: torch.dtype) -> None:
     from tileops.ops.elementwise import AlibiFwdOp
 
-    op = AlibiFwdOp(seq_len=seq_len, num_heads=num_heads, dtype=dtype)
+    op = AlibiFwdOp(seq_len=seq_len, num_heads=num_heads, out_dtype=dtype)
     out = op()
 
     # Reference: slope_h = 2^(-8*(h+1)/H), bias = -slope * |i - j|
@@ -296,7 +296,7 @@ class SinusoidalFixture(FixtureBase):
 def test_sinusoidal(seq_len: int, d_model: int, dtype: torch.dtype) -> None:
     from tileops.ops.elementwise import SinusoidalFwdOp
 
-    op = SinusoidalFwdOp(seq_len=seq_len, d_model=d_model, dtype=dtype)
+    op = SinusoidalFwdOp(seq_len=seq_len, d_model=d_model, out_dtype=dtype)
     out = op()
 
     # Reference
